# docs/swagger.yaml
basePath: /directory/api/v1
definitions:
  types.AuthRequest:
    properties:
      email_address:
        type: string
      otp:
        type: string
    required:
    - email_address
    - otp
    type: object
  types.EmailRequest:
    properties:
      email_address:
        type: string
    required:
    - email_address
    type: object
  types.JWTResponse:
    properties:
      access_token:
        type: string
      profile:
        $ref: '#/definitions/types.UserProfile'
    type: object
  types.Registration:
    properties:
      display_name:
        type: string
      email_address:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    required:
    - email_address
    - first_name
    - last_name
    type: object
  types.UserProfile:
    properties:
      display_name:
        type: string
      email_address:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      last_login:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for business directory api..
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Business directory API
  version: "1.0"
paths:
  /auth/activate:
    get:
      consumes:
      - application/json
      description: Request to verification token
      parameters:
      - description: token
        in: query
        name: token
        type: string
      - description: uid
        in: query
        name: uid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      tags:
      - Auth
  /auth/authentication:
    post:
      consumes:
      - application/json
      description: Request to authentication token
      parameters:
      - description: Authenticate existing user
        in: body
        name: default
        required: true
        schema:
          $ref: '#/definitions/types.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: request to authentication token
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get jwt token
      parameters:
      - description: Login to account
        in: body
        name: default
        required: true
        schema:
          $ref: '#/definitions/types.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                Data:
                  $ref: '#/definitions/types.JWTResponse'
              type: object
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh authorization token
      parameters:
      - description: token
        in: body
        name: default
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      summary: Refresh authorization token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Add a new user
        in: body
        name: default
        required: true
        schema:
          $ref: '#/definitions/types.Registration'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
      tags:
      - Auth
securityDefinitions:
  ApiKeyAuth:
    type: basic
swagger: "2.0"